+++ src/ui/lib.rs
@@ -3,7 +3,9 @@
 //! ui lib
 
 mod client;
+mod history;
 mod kiosk;
 
-pub use client::FeedClient;
-pub use kiosk::{FeedState, FlatFeedState, Kiosk};
+pub use self::client::FeedClient;
+pub use self::history::History;
+pub use self::kiosk::{FeedState, FlatFeedState, Kiosk};